requirements_backup.txt
appdirs==1.4.4
apturl==0.5.2
attrs==21.2.0
bcrypt==3.2.0
beniget==0.4.1
blinker==1.9.0
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.3.1
colorama==0.4.4
cpuset==1.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
easyocr==1.7.2
fasteners==0.14.1
filelock==3.20.0
filetype==1.2.0
Flask==3.1.3
fonttools==4.29.1
fs==2.4.12
fsspec==2025.12.0
future==0.18.2
gast==0.5.2
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
jetson-stats==4.3.2
Jetson.GPIO==2.1.7
Jinja2==3.1.6
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.3
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
packaging==21.3
pandas==1.3.5
paramiko==2.9.3
pexpect==4.8.0
pi_heif==1.2.1
pillow==12.1.1
pillow-avif-plugin==1.5.5
ply==3.11
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.8.1
python-dbusmock==0.27.5
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==1.0.0
roboflow==1.2.14
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
smbus2==0.6.0
sympy==1.14.0
systemd-python==234
tensorrt==10.3.0
tensorrt_dispatch==10.3.0
tensorrt_lean==10.3.0
torchaudio==2.10.0+cpu
tqdm==4.67.3
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
unicodedata2==14.0.0
urllib3==2.6.3
urwid==2.1.2
wadllib==1.3.6
Werkzeug==3.1.6
xdg==5
xkit==0.0.0
zipp==1.0.0
